+++ test/dummy/app/controllers/dashboards_controller.rb
@@ -1,5 +1,6 @@
 class DashboardsController < ApplicationController
   def show
     @snapshot = Janela::Snapshot.order(:taken_at).last
+    @frame = Janela::Frame.order(:id).first
   end
 end