+++ test/dummy/app/controllers/frames_controller.rb
@@ -1,0 +1,5 @@
+class FramesController < ApplicationController
+  def show
+    @frame = Janela::Frame.find(params[:id])
+  end
+end