+++ test/dummy/config/routes.rb
@@ -3,5 +3,6 @@ Rails.application.routes.draw do
   root "dashboards#show"
   get "version", to: "version#show"
   resources :snapshots, only: :show
+  resources :frames, only: :show
   mount Janela::Engine => "/dashboards"
 end